+++ NEWS.md
@@ -46,6 +46,8 @@
 
 * Fixed issues with plotting ordinal or multinomial models.
 
+* Fixed issues with contrasting slopes when `backend` was `"emmeans"`.
+
 # modelbased 0.9.0
 
 ## Breaking Changes